kike suggestions solved v2 :)
