+++ react/tic-tac-toe/src/components/Game/components/Board/components/Square/index.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import styles from './styles.module.scss';
