+++ react/tic-tac-toe/src/components/Game/components/Board/styles.module.scss
@@ -1,4 +1,4 @@
-.boardRow:after {
+.boardRow::after {
   clear: both;
   content: "";
   display: table;
