+++ react/tic-tac-toe/src/components/Game/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-
 import Board from './components/Board';
 import styles from './styles.module.scss';
+import { calculateWinner } from '../../utils/utils';
 
 class Game extends Component {
   state = {
@@ -16,6 +16,7 @@ class Game extends Component {
     const history = this.state.history.slice(0, this.state.stepNumber + 1);
     const current = history[history.length - 1];
     const squares = {...current.squares};
+    
     if (calculateWinner(squares) || squares[i]) {
       return;
     }
@@ -36,12 +37,8 @@ class Game extends Component {
     });
   }
 
-  render() {
-    const history = this.state.history;
-    const current = history[this.state.stepNumber];
-    const winner = calculateWinner(current.squares);
-
-    const moves = history.map((step, move) => {
+  movesMapping(history){
+    return history.map((step, move) => {
       const desc = move ? 'Go to move # ' + move : 'Go to game start';
       return (
         <li key={move}>
@@ -49,6 +46,12 @@ class Game extends Component {
         </li>
       )
     });
+  }
+  
+  render() {
+    const history = this.state.history;
+    const current = history[this.state.stepNumber];
+    const winner = calculateWinner(current.squares);
 
     let status;
     if (winner) 
@@ -66,31 +69,11 @@ class Game extends Component {
         </div>
         <div className={styles.gameInfo}>
           <div>{status}</div>
-          <ol>{moves}</ol>
+          <ol>{this.movesMapping(history)}</ol>
         </div>
       </div>
     );
   }
-}
-
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
 }
 
 export default Game;
